ejercicios_practica/requirements.txt
Django==3.2.2
requests==2.25.1
djangorestframework==3.12.4
django-filter==2.4.0
# Adaptador de la base de datos relacional PostgreSQL para Python.
psycopg2-binary>=2.8
# Toolkit que permite aprovechar al máximo el uso interactivo de Python.
# Muy útil cuando utilizemos el shell de Django.
ipython==8.8.0